windows build: Move platform-dependent files to buildwin
The files here belongs in buildwin. This is partly to keep the windows
specific stuff in one place but also about keeping the root folder
clean.

While on it, use the %~dp0 construct which expands to the batch file
installation part rather than actual working directory to compute base
paths. This is IMHO probably both simpler and more robust.